refactor(measure): extract shared zone-segmentation helper from MeasureSymmetricZones
Relocate the colony-ness -> zone-radii pipeline (dataclass, params, 15 pure
staticmethods, and the compute body) out of MeasureSymmetricZones into a new
side-effect-free _zone_segmentation module. MeasureSymmetricZones._compute_intermediates
becomes a thin delegator to compute_zone_segmentation; a back-compat
_SymmetryIntermediates = ZoneSegmentation alias keeps the inspect/_operate/overlay
helpers untouched. Adds a plate-frame centroid_global field on ZoneSegmentation.

Also repoints refine/_trim_asymmetry.py (an external caller of five of the
relocated staticmethods) to the new module-level functions. Output of
MeasureSymmetricZones is byte-identical (golden regression guard passes).
